Update index.html
How to Use the Square Highlight Feature
Select the text or number you want to highlight with a square

Click the "Square Highlight" button (yellow button)

A black square will appear around your selected text

This feature is perfect for marking text that needs to be cut out or highlighted for special attention. The square will be visible both on screen and when printed.